fixup! fixup! fixup! Factory loading flow refactoring
