+++ packages/dashboard-frontend/src/containers/Loader/Workspace/Steps/Initialize/index.tsx
@@ -23,7 +23,7 @@ import { Workspace } from '../../../../../services/workspace-adapter';
 import { DevWorkspaceStatus } from '../../../../../services/helpers/types';
 import { DisposableCollection } from '../../../../../services/helpers/disposable';
 import { delay } from '../../../../../services/helpers/delay';
-import { MIN_STEP_DURATION_MS, TIMEOUT_TO_STOP_SEC } from '../consts';
+import { MIN_STEP_DURATION_MS, TIMEOUT_TO_STOP_SEC } from '../const';
 import findTargetWorkspace from '../../findTargetWorkspace';
 import { AbstractLoaderStep, LoaderStepProps, LoaderStepState } from '../../../AbstractStep';
