+++ packages/dashboard-frontend/src/containers/Loader/Workspace/Steps/OpenWorkspace/__tests__/index.spec.tsx
@@ -18,7 +18,7 @@ import { screen, waitFor, within } from '@testing-library/react';
 import { WorkspaceParams } from '../../../../../../Routes/routes';
 import { FakeStoreBuilder } from '../../../../../../store/__mocks__/storeBuilder';
 import { DevWorkspaceBuilder } from '../../../../../../store/__mocks__/devWorkspaceBuilder';
-import { TIMEOUT_TO_GET_URL_SEC } from '../../consts';
+import { TIMEOUT_TO_GET_URL_SEC } from '../../const';
 import { LoadingStep } from '../../../../../../components/Loader/Step';
 import { getIdeLoadingSteps } from '../../../../../../components/Loader/Step/buildSteps';
 import getComponentRenderer from '../../../../../../services/__mocks__/getComponentRenderer';
