+++ packages/dashboard-frontend/src/containers/Loader/Workspace/Steps/OpenWorkspace/index.tsx
@@ -21,7 +21,7 @@ import { buildLoaderSteps } from '../../../../../components/Loader/Step/buildSte
 import { WorkspaceLoaderPage } from '../../../../../pages/Loader/Workspace';
 import { DisposableCollection } from '../../../../../services/helpers/disposable';
 import { delay } from '../../../../../services/helpers/delay';
-import { MIN_STEP_DURATION_MS, TIMEOUT_TO_GET_URL_SEC } from '../consts';
+import { MIN_STEP_DURATION_MS, TIMEOUT_TO_GET_URL_SEC } from '../const';
 import findTargetWorkspace from '../../findTargetWorkspace';
 import { Workspace } from '../../../../../services/workspace-adapter';
 import { AbstractLoaderStep, LoaderStepProps, LoaderStepState } from '../../../AbstractStep';
