+++ packages/dashboard-frontend/src/containers/Loader/Workspace/Steps/StartWorkspace/__tests__/index.spec.tsx
@@ -22,7 +22,7 @@ import { ActionCreators } from '../../../../../../store/Workspaces';
 import { AppThunk } from '../../../../../../store';
 import { LoadingStep } from '../../../../../../components/Loader/Step';
 import { getIdeLoadingSteps } from '../../../../../../components/Loader/Step/buildSteps';
-import { TIMEOUT_TO_RUN_SEC } from '../../consts';
+import { TIMEOUT_TO_RUN_SEC } from '../../const';
 import getComponentRenderer from '../../../../../../services/__mocks__/getComponentRenderer';
 import StepStartWorkspace from '..';
