+++ packages/dashboard-frontend/src/containers/Loader/Workspace/Steps/StartWorkspace/index.tsx
@@ -24,7 +24,7 @@ import { DevWorkspaceStatus } from '../../../../../services/helpers/types';
 import { DisposableCollection } from '../../../../../services/helpers/disposable';
 import { delay } from '../../../../../services/helpers/delay';
 import { filterErrorLogs } from '../../../../../services/helpers/filterErrorLogs';
-import { MIN_STEP_DURATION_MS, TIMEOUT_TO_RUN_SEC } from '../consts';
+import { MIN_STEP_DURATION_MS, TIMEOUT_TO_RUN_SEC } from '../const';
 import findTargetWorkspace from '../../findTargetWorkspace';
 import workspaceStatusIs from '../workspaceStatusIs';
 import { Workspace } from '../../../../../services/workspace-adapter';
