+++ packages/dashboard-frontend/src/store/Workspaces/devWorkspaces/index.ts
@@ -78,6 +78,7 @@ interface UpdateWorkspaceStatusAction extends Action {
   workspaceUID: string;
   status: string;
   message: string;
+  mainUrl?: string;
 }
 
 interface UpdateWorkspacesLogsAction extends Action {
@@ -674,6 +675,7 @@ export const reducer: Reducer<State> = (state: State | undefined, action: KnownA
           }
           nextWorkspace.status.phase = action.status;
           nextWorkspace.status.message = action.message;
+          nextWorkspace.status.mainUrl = action.mainUrl;
           return nextWorkspace;
         }),
       });
@@ -724,16 +726,14 @@ export const reducer: Reducer<State> = (state: State | undefined, action: KnownA
 
 async function onStatusUpdateReceived(
   dispatch: ThunkDispatch<AppState, unknown, KnownAction>,
-  statusUpdate: IStatusUpdate & {
-    namespace?: string;
-    workspaceId?: string;
-  },
+  statusUpdate: IStatusUpdate,
 ) {
-  const { status, message, prevStatus, workspaceUID, namespace, workspaceId } = statusUpdate;
+  const { status, message, prevStatus, workspaceUID, namespace, workspaceId, mainUrl } =
+    statusUpdate;
 
   if (status !== prevStatus) {
     const type = 'UPDATE_DEVWORKSPACE_STATUS';
-    dispatch({ type, workspaceUID, message, status });
+    dispatch({ type, workspaceUID, message, status, mainUrl });
 
     const onChangeCallback = onStatusChangeCallbacks.get(workspaceUID);
     if (onChangeCallback) {
